+++ source_gen/source_gen/lib/source_gen.dart
@@ -12,3 +12,4 @@ export 'src/generator_for_annotation.dart';
 export 'src/library.dart' show AnnotatedElement, LibraryReader;
 export 'src/span_for_element.dart' show spanForElement;
 export 'src/type_checker.dart' show TypeChecker;
+export 'src/utils.dart' show typeNameOf;